Use base classes for CMS apphooks
